switch to class
