+++ examples/plain_game/plain_wm_graph.lp
