+++ pyproject.toml
@@ -14,7 +14,7 @@ classifiers = [
   "Programming Language :: Python :: 3.9",
   "Programming Language :: Python :: 3.10",
   "Programming Language :: Python :: 3.11",
-  "Topic :: Scientific/Engineering" :: Computer,
+  "Topic :: Scientific/Engineering :: Information Analysis"
 ]
 dependencies = ["jupyterlab>4.0", "clingo>5.6"]
 description = "A Reasoning Tool for Argumentation Framework"
