+++ src/gamearg/__init__.py
@@ -1,0 +1,4 @@
+"""A Reasoning Tool for Argumentation Framework"""
+__version__ = "0.0.1a"
+
+from .gamearg import *
